[JT-B-J-011] Some changes have been made to the .yml configuration file. Changes have been made to the token generation: new embedded fields firstName, lastName have been added. Added dependencies for working with hibernate (automatic creation of tables in the database), replaced instead of flyway.
A DTO model has been created in which all the fields of a regular user are present, except for the "password" field.

Changes have been made to the controller: user registration methods have been fixed (now the returned object is a DTO model without the "password" field), the method of obtaining user information by ID has been fixed, and the mapping of the method has been fixed (now we get a token at the entrance and get a complete summary of user data from it), the method of obtaining all users has been changed (now as a response, the DTO model is returned, in which the "password" field is missing), the method of changing the user's password has been changed, as well as its mapping (now we receive a token and a new password at the entrance, if the token is valid - we change the password to a new one), the method of deleting the user has been changed, as well as its mapping (now we get a token at the entrance, with which we delete the user).

Changes have been made to the service: the function of converting the basic user model to the DTO model has been added, changes have been made to the ChangePassword function (now we use the new password that came in the request as the encoded password).